Fix debris removal loop implementation
Removal of directories in deep hierarchies was unreliable due to the
fact that `Path.glob()` wouldn't automatically sort file system objects
in a way that makes sure they don't conflict when we attempt to delete
one of them.

We now sort file system objects in *reverse order* and first delete
*all files* before removing directories. This way we make sure that
the directories that are deepest down in the hierarchy are empty when
we attempt to remove them.